% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_code.R
\name{get_table_code}
\alias{get_table_code}
\title{get_table_code}
\usage{
get_table_code(out_tbl, tbl_name, is_console, context)
}
\arguments{
\item{out_tbl}{`data.frame` or `string`}

\item{tbl_name}{`character` named to be used for the table}

\item{is_console}{`logical` if TRUE, the insert_table function was called
from the console, otherwise from an Rmd file using the addin}

\item{context}{context of the call (tells if from console or file, and if
from file allows to retrieve  the lines, etcetera)}
}
\value{
returns the code needed to generate the table, either by creating
 new lines in the Rmd, or by printing it to the console (if is.console = TRUE)
}
\description{
Accessory function used to generate the code needed to generate
 the table in the selcted otuput format
}
\author{
Lorenzo Busetto, phD (2017) <lbusett@gmail.com>
}
